add final update

+++ lib/features/Course/presentation/widgets/video_hours_widget.dart
@@ -5,7 +5,7 @@ import 'package:flutter_screenutil/flutter_screenutil.dart';
 
 class VideoHoursWidget extends StatelessWidget {
   final int videoCount;
-  final int hoursCount;
+  final double hoursCount;
 
   const VideoHoursWidget({
     super.key,
